Refactor package structure
