+++ src/main/java/com/helpunker/assignment/repository/AssignmentRepository.java
@@ -1,0 +1,11 @@
+package com.helpunker.assignment.repository;
+
+import com.helpunker.assignment.Assignment;
+import java.util.Optional;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AssignmentRepository extends JpaRepository<Assignment, UUID> {
+
+    Optional<Assignment> findByRequestId(UUID requestId);
+}
