+++ src/main/java/com/helpunker/helprequest/domain/RequestStatus.java
@@ -1,0 +1,9 @@
+package com.helpunker.helprequest.domain;
+
+public enum RequestStatus {
+    OPEN,
+    ASSIGNED,
+    ONGOING,
+    COMPLETED,
+    CANCELLED
+}
