+++ src/main/java/com/helpunker/helprequest/dto/response/RequestPhotoResponse.java
@@ -1,0 +1,6 @@
+package com.helpunker.helprequest.dto.response;
+
+import java.util.UUID;
+
+public record RequestPhotoResponse(UUID id, String url, String contentType) {
+}
