+++ src/main/java/com/helpunker/helprequest/repository/HelpRequestRepository.java
@@ -1,0 +1,10 @@
+package com.helpunker.helprequest.repository;
+
+import com.helpunker.helprequest.domain.HelpRequest;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+public interface HelpRequestRepository extends JpaRepository<HelpRequest, UUID>,
+        JpaSpecificationExecutor<HelpRequest> {
+}
