+++ src/main/java/com/helpunker/helprequest/service/HelpRequestSearchCriteria.java
@@ -1,0 +1,8 @@
+package com.helpunker.helprequest.service;
+
+import com.helpunker.helprequest.domain.RequestStatus;
+import java.util.UUID;
+
+public record HelpRequestSearchCriteria(
+        RequestStatus status, UUID elderlyId, Double latitude, Double longitude, Double radiusKm) {
+}
