+++ src/main/java/com/helpunker/helprequest/sse/RequestEvent.java
@@ -1,0 +1,6 @@
+package com.helpunker.helprequest.sse;
+
+import com.helpunker.helprequest.dto.response.HelpRequestResponse;
+
+public record RequestEvent(RequestEventType type, HelpRequestResponse payload) {
+}
