+++ src/main/java/com/helpunker/helprequest/sse/RequestEventType.java
@@ -1,0 +1,7 @@
+package com.helpunker.helprequest.sse;
+
+public enum RequestEventType {
+    REQUEST_CREATED,
+    REQUEST_UPDATED,
+    REQUEST_CANCELLED
+}
