--- src/main/java/com/helpunker/modules/assignment/domain/model/Assignment.java
@@ -1,59 +1,0 @@
-package com.helpunker.modules.assignment.domain.model;
-
-import com.helpunker.modules.helprequest.domain.model.HelpRequest;
-import com.helpunker.modules.user.domain.model.User;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.Version;
-import java.time.OffsetDateTime;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "assignments")
-public class Assignment {
-
-    @Id
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "request_id", nullable = false)
-    private HelpRequest request;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "volunteer_id", nullable = false)
-    private User volunteer;
-
-    @Column(name = "accepted_at")
-    private OffsetDateTime acceptedAt;
-
-    @Column(name = "arrived_at")
-    private OffsetDateTime arrivedAt;
-
-    @Column(name = "completed_at")
-    private OffsetDateTime completedAt;
-
-    @Column(name = "cancelled_at")
-    private OffsetDateTime cancelledAt;
-
-    @Column(name = "created_at", nullable = false, insertable = false, updatable = false)
-    private OffsetDateTime createdAt;
-
-    @Version
-    private long version;
-}
