--- src/main/java/com/helpunker/modules/helprequest/infrastructure/sse/RequestEventType.java
@@ -1,7 +1,0 @@
-package com.helpunker.modules.helprequest.infrastructure.sse;
-
-public enum RequestEventType {
-    REQUEST_CREATED,
-    REQUEST_UPDATED,
-    REQUEST_CANCELLED
-}
