--- src/main/java/com/helpunker/modules/outbox/domain/repository/OutboxEventRepository.java
@@ -1,7 +1,0 @@
-package com.helpunker.modules.outbox.domain.repository;
-
-import com.helpunker.modules.outbox.domain.model.OutboxEvent;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OutboxEventRepository extends JpaRepository<OutboxEvent, Long> {
-}
