+++ src/main/java/com/helpunker/outbox/repository/OutboxEventRepository.java
@@ -1,0 +1,7 @@
+package com.helpunker.outbox.repository;
+
+import com.helpunker.outbox.OutboxEvent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OutboxEventRepository extends JpaRepository<OutboxEvent, Long> {
+}
