+++ src/main/java/com/helpunker/user/UserRole.java
@@ -1,0 +1,7 @@
+package com.helpunker.user;
+
+public enum UserRole {
+    ELDERLY,
+    VOLUNTEER,
+    ADMIN
+}
