+++ src/main/java/com/helpunker/user/repository/DeviceTokenRepository.java
@@ -1,0 +1,11 @@
+package com.helpunker.user.repository;
+
+import com.helpunker.user.DeviceToken;
+import java.util.List;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DeviceTokenRepository extends JpaRepository<DeviceToken, UUID> {
+
+    List<DeviceToken> findByUserId(UUID userId);
+}
