+++ src/main/java/com/helpunker/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.helpunker.user.repository;
+
+import com.helpunker.user.User;
+import java.util.Optional;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+    Optional<User> findByEmail(String email);
+
+    Optional<User> findByPhone(String phone);
+}
